requirements/staging.txt
#contains packages for staging/testing area.
pytest
pytest-cov
pytest-django